wip: arm64: dts: qcom: x1e80100-crd: Add WiFi/BT pwrseq
Add the WiFi/BT nodes for CRD and describe the regulators for the WCN7850
combo chip using the new power sequencing bindings.

Unlike QCP, CRD has the WCN7850 chip on a removable M.2 card. In this case,
there is only a standard 3.3V and 1.8V supply that we need to turn on (plus
the two enable GPIOs). The other voltages (0.95V and 1.9V) are derived on
the M.2 card, by regulators that are not part of the CRD mainboard.

Right now the bindings require describing all of these regulators. We need
new bindings to describe these M.2 cards with only the two actual supplies.
For now just copy the necessary additions from QCP. Functionally this works
correctly for CRD, because from a hardware perspective we need to do
exactly the same for both devices (toggle the GPIO to get the supplies up).

Signed-off-by: Stephan Gerhold <stephan.gerhold@linaro.org>
[ johan: rebase on x1-crd.dtsi split ]
Signed-off-by: Johan Hovold <johan+linaro@kernel.org>
Signed-off-by: Neil Armstrong <neil.armstrong@linaro.org>
Signed-off-by: Abel Vesa <abel.vesa@linaro.org>