media: platform: qcom/iris: add sm8650 support
Add support for the SM8650 platform by re-using the SM8550
definitions and using the vpu33 ops.

Move the reset tables that diffes in a per-SoC platform
header, that will contain mode SoC specific data when
more codecs are introduced.

The SM8650/vpu33 requires more reset lines, but the H.264
decoder capabilities are identical.

Tested-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org> # x1e Dell
Signed-off-by: Neil Armstrong <neil.armstrong@linaro.org>