media: iris: Skip flush on first sequence change
Add a condition to skip the flush operation during the first sequence
change event. At this point, the capture queue is not streaming, making
the flush unnecessary.

Additionally, remove the reinit_completion call for the flush completion
signal, as it is not needed. This simplifies the code and avoids
unnecessary reinitialization.

Acked-by: Vikash Garodia <quic_vgarodia@quicinc.com>
Signed-off-by: Dikshita Agarwal <quic_dikshita@quicinc.com>
Link: https://lore.kernel.org/r/20250502-qcom-iris-hevc-vp9-v3-16-552158a10a7d@quicinc.com
Signed-off-by: Neil Armstrong <neil.armstrong@linaro.org>